// packages/appsync-local-codegen-plugin/src/visitors/appsync-visitor.ts
export type CodeGenDirective = {
  name: string;
  arguments: CodeGenArgumentsMap;
};
export type CodeGenDirectives = CodeGenDirective[];
export type CodeGenField = TypeInfo & {
  name: string;
  directives: CodeGenDirectives;
};
export type TypeInfo = {
  type: string;
  isList: boolean;
  isNullable: boolean;
  baseType?: GraphQLNamedType | null;
};
export type CodeGenModel = {
  name: string;
  type: 'model';
  directives: CodeGenDirectives;
  fields: CodeGenField[];
};